requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart 
authlib 
fastapi[all]